feat(engine): delay job header encoding
We want to add additional headers to the customHeaders property of
JobRecord. These headers will only be available at job creation-time. So
we need to delay encoding the headers to the moment when the job CREATED
record is written.

The logic to encode the headers was also duplicated between the
UserTaskTransformer and the JobWorkerElementTransformer. This commit
removes that duplication.