feat: automatically add copyright header upon creating a new file
These settings in intellij will automatically add the correct copyright header to a new file that's created. By default it will add the Zeebe 1.1 license header. There are 5 exceptions to this:

1. zeebe-java-client
2. zeebe-exporter-api
3. zeebe-protocol
4. zeebe-gateway-protocol-impl
5. zeebe-bpmn-model

For these 5 modules we will automatically add the Apache 2.0 license header.

The Go client should also be included in this list, however IntelliJ scopes only support the addition of packages. Since these packages are a Java concept we cannot include the Go files to these exceptions. No copyright header will be added automatically for new files in the go client.